Fix markdownlint (#81)
* Fix for Markdownlint

We added markdownlint checks, but never updated the
README to reflect this. This commit makes a large
number of changes to the README to try to make
markdownlint happy. The DEI (BEAR) info was obsolete
so I removed it.

For example, Markdown permits unordered lists to begin
with either * or -, but Markdownlint wants you to pick
one notation and be consistent. I decided to use
asterisk for unordered lists, since they look more
like bullets.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update README

Update the README to better reflect reality. For example,
stream 1 shouldn't be emphasized, DEI/BEAR is its own thing,
and add a link to 2025 meeting notes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

---------

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>